+++ bin/reth/Cargo.toml
@@ -20,6 +20,7 @@ reth-db = { workspace = true, features = ["mdbx"] }
 reth-provider = { workspace = true }
 reth-revm.workspace = true
 reth-stages.workspace = true
+reth-stages-api.workspace = true
 reth-interfaces = { workspace = true, features = ["clap"] }
 reth-transaction-pool.workspace = true
 reth-beacon-consensus.workspace = true
