+++ crates/blockchain-tree/Cargo.toml
@@ -16,7 +16,7 @@ reth-primitives.workspace = true
 reth-interfaces.workspace = true
 reth-db.workspace = true
 reth-provider.workspace = true
-reth-stages.workspace = true
+reth-stages-api.workspace = true
 reth-trie = { workspace = true, features = ["metrics"] }
 reth-trie-parallel = { workspace = true, features = ["parallel"] }
