+++ crates/consensus/auto-seal/Cargo.toml
@@ -17,7 +17,7 @@ reth-beacon-consensus.workspace = true
 reth-primitives.workspace = true
 reth-interfaces.workspace = true
 reth-provider.workspace = true
-reth-stages.workspace = true
+reth-stages-api.workspace = true
 reth-revm.workspace = true
 reth-transaction-pool.workspace = true
 reth-evm.workspace = true
