+++ crates/node-builder/src/setup.rs
@@ -44,7 +44,7 @@ pub async fn build_networked_pipeline<DB, Client, EvmConfig>(
     consensus: Arc<dyn Consensus>,
     provider_factory: ProviderFactory<DB>,
     task_executor: &TaskExecutor,
-    metrics_tx: reth_stages::MetricEventsSender,
+    metrics_tx: reth_stages_api::MetricEventsSender,
     prune_config: Option<PruneConfig>,
     max_block: Option<BlockNumber>,
     static_file_producer: StaticFileProducer<DB>,
@@ -92,7 +92,7 @@ pub async fn build_pipeline<DB, H, B, EvmConfig>(
     body_downloader: B,
     consensus: Arc<dyn Consensus>,
     max_block: Option<u64>,
-    metrics_tx: reth_stages::MetricEventsSender,
+    metrics_tx: reth_stages_api::MetricEventsSender,
     prune_config: Option<PruneConfig>,
     static_file_producer: StaticFileProducer<DB>,
     evm_config: EvmConfig,
