+++ crates/prune/Cargo.toml
@@ -34,7 +34,8 @@ tokio-stream.workspace = true
 [dev-dependencies]
 # reth
 reth-db = { workspace = true, features = ["test-utils"] }
-reth-stages = { workspace = true, features = ["test-utils"] }
+#reth-stages-api = { workspace = true, features = ["test-utils"] }
+reth-stages = { workspace = true}
 reth-tracing.workspace = true
 
 # misc
