+++ crates/stages-api/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "reth-stages-api"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+exclude.workspace = true
+
+[dependencies]
+reth-primitives.workspace = true
+reth-provider.workspace = true
+reth-db.workspace = true
+reth-interfaces.workspace = true
+
+reth-metrics.workspace = true
+metrics.workspace = true
+
+thiserror.workspace = true
+tokio.workspace = true
+tracing.workspace = true
+
+[lints]
+workspace = true
