+++ crates/stages-api/src/lib.rs
@@ -1,0 +1,10 @@
+//! Staged syncing primitives for reth.
+mod error;
+mod metrics;
+mod pipeline;
+mod stage;
+
+pub use crate::metrics::*;
+pub use error::*;
+pub use pipeline::*;
+pub use stage::*;
