+++ crates/stages-api/src/metrics/listener.rs
@@ -45,7 +45,7 @@ pub enum MetricEvent {
 #[derive(Debug)]
 pub struct MetricsListener {
     events_rx: UnboundedReceiver<MetricEvent>,
-    pub(crate) sync_metrics: SyncMetrics,
+    pub sync_metrics: SyncMetrics,
 }
 
 impl MetricsListener {
