+++ crates/stages-api/src/metrics/sync_metrics.rs
@@ -6,14 +6,14 @@ use reth_primitives::stage::StageId;
 use std::collections::HashMap;
 
 #[derive(Debug, Default)]
-pub(crate) struct SyncMetrics {
-    pub(crate) stages: HashMap<StageId, StageMetrics>,
-    pub(crate) execution_stage: ExecutionStageMetrics,
+pub struct SyncMetrics {
+    pub stages: HashMap<StageId, StageMetrics>,
+    pub execution_stage: ExecutionStageMetrics,
 }
 
 impl SyncMetrics {
     /// Returns existing or initializes a new instance of [StageMetrics] for the provided [StageId].
-    pub(crate) fn get_stage_metrics(&mut self, stage_id: StageId) -> &mut StageMetrics {
+    pub fn get_stage_metrics(&mut self, stage_id: StageId) -> &mut StageMetrics {
         self.stages
             .entry(stage_id)
             .or_insert_with(|| StageMetrics::new_with_labels(&[("stage", stage_id.to_string())]))
@@ -22,19 +22,19 @@ impl SyncMetrics {
 
 #[derive(Metrics)]
 #[metrics(scope = "sync")]
-pub(crate) struct StageMetrics {
+pub struct StageMetrics {
     /// The block number of the last commit for a stage.
-    pub(crate) checkpoint: Gauge,
+    pub checkpoint: Gauge,
     /// The number of processed entities of the last commit for a stage, if applicable.
-    pub(crate) entities_processed: Gauge,
+    pub entities_processed: Gauge,
     /// The number of total entities of the last commit for a stage, if applicable.
-    pub(crate) entities_total: Gauge,
+    pub entities_total: Gauge,
 }
 
 /// Execution stage metrics.
 #[derive(Metrics)]
 #[metrics(scope = "sync.execution")]
-pub(crate) struct ExecutionStageMetrics {
+pub struct ExecutionStageMetrics {
     /// The total amount of gas processed (in millions)
-    pub(crate) mgas_processed_total: Counter,
+    pub mgas_processed_total: Counter,
 }
