+++ crates/stages-api/src/pipeline/mod.rs
@@ -1,0 +1,4 @@
+mod ctrl;
+mod event;
+pub use crate::pipeline::ctrl::ControlFlow;
+pub use event::*;
