+++ crates/stages/src/lib.rs
@@ -77,8 +77,6 @@
 )]
 #![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 
-mod error;
-mod metrics;
 mod pipeline;
 mod stage;
 mod util;
@@ -95,7 +93,5 @@ pub mod stages;
 
 pub mod sets;
 
-pub use crate::metrics::*;
-pub use error::*;
 pub use pipeline::*;
 pub use stage::*;
