+++ crates/stages/src/prelude.rs
@@ -1,6 +1,5 @@
 pub use crate::{
-    error::{PipelineError, StageError},
-    pipeline::{Pipeline, PipelineBuilder, PipelineEvent, StageSet, StageSetBuilder},
+    pipeline::{Pipeline, PipelineBuilder, StageSet, StageSetBuilder},
     sets::{
         DefaultStages, ExecutionStages, HashingStages, HistoryIndexingStages, OfflineStages,
         OnlineStages,
