+++ crates/stages/src/stages/headers.rs
@@ -1,4 +1,4 @@
-use crate::{BlockErrorKind, ExecInput, ExecOutput, Stage, StageError, UnwindInput, UnwindOutput};
+use crate::Stage;
 use futures_util::StreamExt;
 use reth_codecs::Compact;
 use reth_config::config::EtlConfig;
@@ -25,6 +25,9 @@ use reth_provider::{
     providers::{StaticFileProvider, StaticFileWriter},
     BlockHashReader, DatabaseProviderRW, HeaderProvider, HeaderSyncGap, HeaderSyncGapProvider,
     HeaderSyncMode,
+};
+use reth_stages_api::{
+    BlockErrorKind, ExecInput, ExecOutput, StageError, UnwindInput, UnwindOutput,
 };
 use std::{
     sync::Arc,
