+++ crates/stages/src/stages/index_account_history.rs
@@ -1,5 +1,5 @@
 use super::{collect_history_indices, load_history_indices};
-use crate::{ExecInput, ExecOutput, Stage, StageError, UnwindInput, UnwindOutput};
+use crate::Stage;
 use reth_config::config::EtlConfig;
 use reth_db::{
     database::Database, models::ShardedKey, table::Decode, tables, transaction::DbTxMut,
@@ -11,6 +11,7 @@ use reth_primitives::{
 use reth_provider::{
     DatabaseProviderRW, HistoryWriter, PruneCheckpointReader, PruneCheckpointWriter,
 };
+use reth_stages_api::{ExecInput, ExecOutput, StageError, UnwindInput, UnwindOutput};
 use std::fmt::Debug;
 use tracing::info;
