+++ crates/stages/src/stages/merkle.rs
@@ -1,4 +1,4 @@
-use crate::{BlockErrorKind, ExecInput, ExecOutput, Stage, StageError, UnwindInput, UnwindOutput};
+use crate::Stage;
 use reth_codecs::Compact;
 use reth_db::{
     database::Database,
@@ -14,6 +14,9 @@ use reth_primitives::{
 use reth_provider::{
     DatabaseProviderRW, HeaderProvider, ProviderError, StageCheckpointReader,
     StageCheckpointWriter, StatsReader,
+};
+use reth_stages_api::{
+    BlockErrorKind, ExecInput, ExecOutput, StageError, UnwindInput, UnwindOutput,
 };
 use reth_trie::{IntermediateStateRootState, StateRoot, StateRootProgress};
 use std::fmt::Debug;
