+++ testing/ef-tests/Cargo.toml
@@ -19,7 +19,7 @@ asm-keccak = ["reth-primitives/asm-keccak"]
 reth-primitives.workspace = true
 reth-db = { workspace = true, features = ["mdbx", "test-utils"] }
 reth-provider = { workspace = true, features = ["test-utils"] }
-reth-stages.workspace = true
+reth-stages-api.workspace = true
 reth-interfaces.workspace = true
 reth-revm.workspace = true
 reth-node-ethereum.workspace = true
